+++ Classes/Item.cs
@@ -6,5 +6,7 @@ namespace VidsNet
 
         public string Path { get; set; } 
         public int Id { get; set; }
+
+        public ItemType Type {get;set;}
     }
 }