+++ changelog.md
@@ -1,5 +1,17 @@
 #VidsNET
 
+##0.1.1
+###Back-End
+- [x] Add BaseScanner class
+- [x] Add VideoScanner class
+- [x] Add ItemType to Item class
+- [x] Add static file serving
+
+###Front-End
+- [x] Add Bootstrap 
+- [x] Add jQuery
+
+
 ##0.1.0
 ###Back-End
 - [x] Add EntityFramework support